+++ mmedit/evaluation/functional/__init__.py
@@ -1,4 +1,10 @@
 # Copyright (c) OpenMMLab. All rights reserved.
+from .fid_inception import InceptionV3
 from .gaussian_funcs import gauss_gradient
+from .inception_utils import (disable_gpu_fuser_on_pt19, load_inception,
+                              prepare_inception_feat, prepare_vgg_feat)
 
-__all__ = ['gauss_gradient']
+__all__ = [
+    'gauss_gradient', 'InceptionV3', 'disable_gpu_fuser_on_pt19',
+    'load_inception', 'prepare_vgg_feat', 'prepare_inception_feat'
+]
