+++ mmedit/evaluation/metrics/fid.py
@@ -10,9 +10,9 @@ from torch import Tensor
 from torch.utils.data.dataloader import DataLoader
 
 from mmedit.registry import METRICS
+from ..functional import (disable_gpu_fuser_on_pt19, load_inception,
+                          prepare_inception_feat)
 from .base_gen_metric import GenerativeMetric
-from .inception_utils import (disable_gpu_fuser_on_pt19, load_inception,
-                              prepare_inception_feat)
 
 
 @METRICS.register_module('FID-Full')
