+++ mmedit/evaluation/metrics/inception_score.py
@@ -12,8 +12,9 @@ from torch import Tensor
 from torch.utils.data.dataloader import DataLoader
 
 from mmedit.registry import METRICS
+# from .inception_utils import disable_gpu_fuser_on_pt19, load_inception
+from ..functional import disable_gpu_fuser_on_pt19, load_inception
 from .base_gen_metric import GenerativeMetric
-from .inception_utils import disable_gpu_fuser_on_pt19, load_inception
 
 
 @METRICS.register_module('IS')
