+++ mmedit/models/data_preprocessors/__init__.py
@@ -1,6 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .batch_process import split_batch, stack_batch
-from .data_processor import EditDataPreprocessor
+from .edit_data_processor import EditDataPreprocessor, split_batch, stack_batch
 from .gen_preprocessor import GenDataPreprocessor
 from .mattor_preprocessor import MattorPreprocessor
